tests/wpt/web-platform-tests/import-maps/acquire-import-maps-flag/worker-request/success.tentative.html

<!DOCTYPE html>
<html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
// https://github.com/WICG/import-maps/blob/master/spec.md#when-import-maps-can-be-encountered
const t = async_test(
  'After module worker creation, import maps should fire error events');
const log = [];
new Worker('../../resources/empty.js?pipe=trickle(d1)', {type: "module"});
</script>
<script type="importmap" onerror="t.done()">
{
  "imports": {
    "../../resources/log.js?pipe=sub&name=A": "../../resources/log.js?pipe=sub&name=B"
  }
}
</script>
<script>
promise_test(() => {
  return import("../../resources/log.js?pipe=sub&name=A")
    .then(() => assert_array_equals(log, ["A"]))
  },
  'After module worker creation import maps are not effective');
</script>
